make crud generator

+++ routes/web.php
@@ -1,16 +1,5 @@
 <?php
 
-/*
-|--------------------------------------------------------------------------
-| Web Routes
-|--------------------------------------------------------------------------
-|
-| Here is where you can register web routes for your application. These
-| routes are loaded by the RouteServiceProvider within a group which
-| contains the "web" middleware group. Now create something great!
-|
-*/
-
 Route::get('/', function () {
     return view('welcome');
 });
@@ -20,11 +9,12 @@ Auth::routes();
 Route::get('/home', 'HomeController@index')->name('home');
 
 //Create Language Translation route
-Route::get('languages', 'LanguageTranslationController@index')->name('languages');
-Route::post('translations/update', 'LanguageTranslationController@transUpdate')->name('translation.update.json');
-Route::post('translations/updateKey', 'LanguageTranslationController@transUpdateKey')->name('translation.update.json.key');
-Route::delete('translations/destroy/{key}', 'LanguageTranslationController@destroy')->name('translations.destroy');
-Route::post('translations/create', 'LanguageTranslationController@store')->name('translations.create');
+Route::get('languages', 'Admin\\LanguageTranslationController@index')->name('languages');
+Route::post('translations/update', 'Admin\\LanguageTranslationController@transUpdate')->name('translation.update.json');
+Route::post('translations/updateKey', 'Admin\\LanguageTranslationController@transUpdateKey')->name('translation.update.json.key');
+Route::delete('translations/destroy/{key}', 'Admin\\LanguageTranslationController@destroy')->name('translations.destroy');
+Route::post('translations/create', 'Admin\\LanguageTranslationController@store')->name('translations.create');
+
 Route::get('check-translation', function(){
 	\App::setLocale('fr');
 	
@@ -34,13 +24,13 @@ Route::get('check-translation', function(){
 // Language Route for Switching
 Route::post('/lang', array(
     'Middleware' => 'LanguageSwitcher',
-    'uses' => 'LanguageController@index',
+    'uses' => 'Admin\\LanguageController@index',
 ))->name('lang');
 
 // For Language direct URL link
 Route::get('/lang/{lang}', array(
     'Middleware' => 'LanguageSwitcher',
-    'uses' => 'LanguageController@change',
+    'uses' => 'Admin\\LanguageController@change',
 ))->name('langChange');
 //End of Language Route
